requirements.txt
click==6.7
futures==3.0.5
google-auth==0.5.0
google-auth-httplib2==0.0.2
google-cloud-core==0.22.1
google-cloud-translate==0.22.0
googleapis-common-protos==1.5.0
httplib2==0.9.2
korean==0.1.8
MarkupSafe==0.23
protobuf==3.1.0.post1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pytz==2016.10
requests==2.13.0
rsa==3.4.2
six==1.10.0
twilio==5.7.0
Werkzeug==0.11.15
wheel==0.24.0
